slothops-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
PyGithub==2.3.0
google-genai
pydantic==2.9.0
aiosqlite==0.20.0
python-dotenv==1.0.1
sse-starlette==2.1.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
